refactor: Users에서 dto, type 정의들을 개별 파일로 분리
현재 한 파일에 여러 개의 정의를 모아두었는데 이 경우 상위 폴더가 복수형인 의미가 없는 것 같아 개별 파일로 분리했다.
