+++ backend/src/users/types/create-user.type.ts
@@ -1,0 +1,9 @@
+export interface CreateUser {
+    oauthNickname: string;
+    email: string;
+    profileImageUrl: string;
+    oauthId: string;
+    oauthAccessToken: string;
+    oauthRefreshToken: string;
+    refreshToken: string;
+}
