+++ backend/src/users/types/update-user.type.ts
@@ -1,0 +1,4 @@
+export interface UpdateUser {
+    nickname: string;
+    profileImageUrl: string;
+}
