+++ backend/src/users/types/user-profile.type.ts
@@ -1,0 +1,8 @@
+import { OauthProvider } from 'src/auth/types/oauth-provider.type';
+
+export interface UserProfile {
+    nickname: string;
+    email: string;
+    profileImageUrl: string;
+    provider: OauthProvider;
+}
